HW3_Python
Somehow when I downloaded the pdf with assignment 3 on it, only the
first page downloaded (which, unfortunately for me, happened to be
spaced in a way that gave no indication there were any other pages to
the assignment), so I had no idea there were questions 8-14 until I got
to class on thursday evening. So my apologies for the tardiness of
those problems.